trying fix init test
