+++ pipelines/stripe_analytics/__init__.py
@@ -1,3 +1,113 @@
 """ This pipeline uses Stripe API and dlt to load data such as Customer, Subscription, Event and etc. to the database and to calculate the MRR and churn rate. """
 
-from .stripe_analytics import stripe_source, metrics_resource
+from enum import Enum
+from typing import Any, Dict, Generator, Optional
+
+import dlt
+import stripe
+from dlt.common import pendulum
+from dlt.extract.source import DltResource
+from pendulum import DateTime
+
+from .metrics import calculate_mrr, churn_rate
+
+
+class Endpoints(Enum):
+    customers: str = "Customer"
+    subscriptions: str = "Subscription"
+    prices: str = "Price"
+    discounts: str = "Coupon"
+    events: str = "Event"
+
+
+def stripe_get_data(
+    resource: Endpoints,
+    start_date: Optional[Any] = None,
+    end_date: Optional[Any] = None,
+    **kwargs: Any,
+) -> Dict[Any, Any]:
+    if start_date:
+        if isinstance(start_date, str):
+            start_date = pendulum.from_format(start_date, "%Y-%m-%dT%H:%M:%SZ")
+        if isinstance(start_date, DateTime):
+            # convert to unix timestamp
+            start_date = int(start_date.timestamp())
+    if end_date:
+        # convert to unix timestamp
+        end_date = int(end_date.timestamp())
+
+    if resource == Endpoints.subscriptions:
+        kwargs.update({"status": "all"})
+
+    resource_dict = getattr(stripe, resource.value).list(
+        created={"gte": start_date, "lt": end_date}, limit=100, **kwargs
+    )
+    return dict(resource_dict)
+
+
+@dlt.source(name="stripe")
+def stripe_source(
+    stripe_secret_key: str = dlt.secrets.value,
+    start_date: Optional[DateTime] = None,
+    end_date: Optional[DateTime] = None,
+) -> Generator[DltResource, Any, None]:
+    stripe.api_key = stripe_secret_key
+    stripe.api_version = "2022-11-15"
+
+    start_date_unix = int(start_date.timestamp()) if start_date is not None else -1
+
+    def get_resource(
+        endpoint: Endpoints,
+        created: Optional[Any] = dlt.sources.incremental(
+            "created", initial_value=start_date_unix
+        ),
+    ) -> Generator[Dict[Any, Any], Any, None]:
+        get_more = True
+        starting_after = None
+        start_value = created.last_value
+
+        while get_more:
+            response = stripe_get_data(
+                endpoint,
+                start_date=start_value,
+                end_date=end_date,
+                starting_after=starting_after,
+            )
+            get_more = response["has_more"]
+
+            if len(response["data"]) > 0:
+                starting_after = response["data"][-1]["id"]
+
+            yield response["data"]
+
+    for endpoint in Endpoints:
+        yield dlt.resource(
+            get_resource,
+            name=endpoint.value,
+            write_disposition="merge",
+            primary_key="id",
+        )(endpoint)
+
+
+@dlt.resource(name="Metrics", write_disposition="append", primary_key="created")
+def metrics_resource() -> Generator[Dict[str, Any], Any, None]:
+    pipeline = dlt.current.pipeline()  # type: ignore
+    with pipeline.sql_client() as client:
+        with client.execute_query("SELECT * FROM subscription") as table:
+            sub_info = table.df()
+
+    # Access to events through the Retrieve Event API is guaranteed only for 30 days.
+    # But we probably have old data in the database.
+    with pipeline.sql_client() as client:
+        with client.execute_query(
+            "SELECT * FROM event WHERE created > %s", pendulum.now().subtract(days=30)
+        ) as table:
+            event_info = table.df()
+
+    mrr = calculate_mrr(sub_info)
+    print(f"MRR: {mrr}")
+
+    churn = churn_rate(event_info, sub_info)
+    print(f"Churn rate: {round(churn * 100, 1)}%")
+
+    yield {"MRR": mrr, "Churn rate": churn, "created": pendulum.now()}
